+++ src/main/java/ru/yandex/practicum/filmorate/dto/UpdateFilmRequest.java
@@ -28,7 +28,7 @@ public class UpdateFilmRequest {
     LocalDate releaseDate;
     @Positive(message = "Продолжительность должна быть положительным числом")
     @NotNull(message = "Продолжительность не может быть пустой")
-    Long duration;
+    Integer duration;
     MpaRating mpaRating;
     Set<Genre> genres = new HashSet<>();
